fix(pov_agent): Corrected last_bid/last_ask type annotations from float to int
- Prices are integer cents per ABIDES convention; Optional[float] was incorrect

+++ abides-markets/abides_markets/agents/pov_execution_agent.py
@@ -121,8 +121,8 @@ class POVExecutionAgent(TradingAgent):
         self.execution_complete: bool = False
 
         # Market data
-        self.last_bid: Optional[float] = None
-        self.last_ask: Optional[float] = None
+        self.last_bid: Optional[int] = None
+        self.last_ask: Optional[int] = None
         self.last_transacted_volume: int = 0
 
         # Logging and statistics
